# Improve performance for xml parsing