--- a/src/Jackalope/Transport/DoctrineDBAL/XmlParser/XmlToPropsParser.php
+++ b/src/Jackalope/Transport/DoctrineDBAL/XmlParser/XmlToPropsParser.php
@@ -53,12 +53,12 @@ class XmlToPropsParser
     /**
      * @var mixed
      */
-    private $currentValueData = null;
+    private $currentValueData = '';
 
     /**
      * @var mixed
      */
-    private $currentPropData = null;
+    private $currentPropData = '';
 
     /**
      * @var \stdClass
@@ -134,50 +134,16 @@ class XmlToPropsParser
             return;
         }
 
-        // it could be that there exist a sv:property node without a sv:value
-        // in this case the value is set on the property data
-        if ($name === 'SV:PROPERTY' && empty($this->currentValues)) {
-            $this->currentValueData = $this->currentPropData;
+        if ($name === 'SV:VALUE') {
+            $this->addCurrentValue($this->currentValueData);
+            $this->currentValueData = '';
         }
 
-        if ($this->currentValueData) {
-            switch ($this->lastPropertyType) {
-                case PropertyType::NAME:
-                case PropertyType::URI:
-                case PropertyType::WEAKREFERENCE:
-                case PropertyType::REFERENCE:
-                case PropertyType::PATH:
-                case PropertyType::DECIMAL:
-                case PropertyType::STRING:
-                    $this->currentValues[] = $this->currentValueData;
-                    break;
-                case PropertyType::BOOLEAN:
-                    $this->currentValues[] = (bool)$this->currentValueData;
-                    break;
-                case PropertyType::LONG:
-                    $this->currentValues[] = (int)$this->currentValueData;
-                    break;
-                case PropertyType::BINARY:
-                    $this->currentValues[] = (int)$this->currentValueData;
-                    break;
-                case PropertyType::DATE:
-                    $date = $this->currentValueData;
-                    if ($date) {
-                        $date = new \DateTime($date);
-                        $date->setTimezone(new \DateTimeZone(date_default_timezone_get()));
-                        // Jackalope expects a string, might make sense to refactor to allow DateTime instances too
-                        $date = $this->valueConverter->convertType($date, PropertyType::STRING);
-                    }
-                    $this->currentValues[] = $date;
-                    break;
-                case PropertyType::DOUBLE:
-                    $this->currentValues[] = (double)$this->currentValueData;
-                    break;
-                default:
-                    throw new \InvalidArgumentException("Type with constant $this->lastPropertyType not found.");
-            }
-
-            $this->currentValueData = null;
+        // it could be that there exist a sv:property node without a sv:value
+        // in this case the value is set on the property data
+        if ($name === 'SV:PROPERTY' && empty($this->currentValues) && !$this->lastPropertyMultiValued) {
+            $this->addCurrentValue($this->currentPropData);
+            $this->currentPropData = '';
         }
 
         if ($name !== 'SV:PROPERTY') {
@@ -194,9 +160,9 @@ class XmlToPropsParser
                 break;
         }
 
-        $this->currentValueData = null;
-        $this->currentPropData = null;
         $this->currentValues = [];
+        $this->currentValueData = '';
+        $this->currentPropData = '';
         $this->lastPropertyName = null;
         $this->lastPropertyType = null;
         $this->lastPropertyMultiValued = null;
@@ -209,15 +175,54 @@ class XmlToPropsParser
         }
 
         if ($this->currentTag === 'SV:VALUE') {
-            $this->currentValueData = $data;
+            $this->currentValueData .= $data;
 
             return;
         }
 
         if ($this->currentTag === 'SV:PROPERTY') {
-            $this->currentPropData = $data;
+            $this->currentPropData .= $data;
 
             return;
+        }
+    }
+
+    private function addCurrentValue($value): void
+    {
+        switch ($this->lastPropertyType) {
+            case PropertyType::NAME:
+            case PropertyType::URI:
+            case PropertyType::WEAKREFERENCE:
+            case PropertyType::REFERENCE:
+            case PropertyType::PATH:
+            case PropertyType::DECIMAL:
+            case PropertyType::STRING:
+                $this->currentValues[] = (string)$value;
+                break;
+            case PropertyType::BOOLEAN:
+                $this->currentValues[] = (bool)$value;
+                break;
+            case PropertyType::LONG:
+                $this->currentValues[] = (int)$value;
+                break;
+            case PropertyType::BINARY:
+                $this->currentValues[] = (int)$value;
+                break;
+            case PropertyType::DATE:
+                $date = $value;
+                if ($date) {
+                    $date = new \DateTime($date);
+                    $date->setTimezone(new \DateTimeZone(date_default_timezone_get()));
+                    // Jackalope expects a string, might make sense to refactor to allow DateTime instances too
+                    $date = $this->valueConverter->convertType($date, PropertyType::STRING);
+                }
+                $this->currentValues[] = $date;
+                break;
+            case PropertyType::DOUBLE:
+                $this->currentValues[] = (double)$value;
+                break;
+            default:
+                throw new \InvalidArgumentException("Type with constant $this->lastPropertyType not found.");
         }
     }
 }