… tests/Jackalope/Transport/DoctrineDBAL/XmlParser/XmlToPropsParserTest.php