I currently did analyse the performance of @sulu in https://github.com/sulu/sulu/pull/6089 after some optimizations on our side most time is spend now in `xmlToColumns` see the following benchmark:

## Before

Response time ~1080ms

 - https://blackfire.io/profiles/af73cb80-e327-4d12-9e76-d729b7097b86/graph?settings%5Bdimension%5D=wt&settings%5Bdisplay%5D=landscape&settings%5BtabPane%5D=nodes&selected=&callname=main()
 
#### Apache Bench Test:
 
 ```bash
ab -n 50 -c 1 -l http://localhost/my-uri
 ```
 
 <details>
 <summary>Result: Requests per second:    1.01 [#/sec] (mean)</summary>
 
 ```bash
 Server Software:        nginx/1.21.0
Server Hostname:        stern-nl.localhost
Server Port:            80

Document Path:          /my-uri
Document Length:        348158 bytes

Concurrency Level:      1
Time taken for tests:   67.569 seconds
Complete requests:      50
Failed requests:        0
Total transferred:      17429509 bytes
HTML transferred:       17407759 bytes
Requests per second:    0.74 [#/sec] (mean)
Time per request:       1351.384 [ms] (mean)
Time per request:       1351.384 [ms] (mean, across all concurrent requests)
Transfer rate:          251.90 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:   974 1351 479.1   1156    3550
Waiting:      973 1344 474.9   1155    3548
Total:        974 1351 479.1   1156    3550

Percentage of the requests served within a certain time (ms)
  50%   1156
  66%   1399
  75%   1533
  80%   1590
  90%   1896
  95%   2427
  98%   3550
  99%   3550
 100%   3550 (longest request)
 ```
 
 </details>
 
<img width="626" alt="Bildschirmfoto 2021-06-01 um 01 36 17" src="https://user-images.githubusercontent.com/1698337/120249034-c9f46300-c279-11eb-84a4-89a7a858cf37.png">

 
## After

Response time ~580ms
 
#### Apache Bench Test:
 
 ```bash
ab -n 50 -c 1 -l http://localhost/my-uri
 ```
 
 <details>
 <summary>Result: Requests per second:    1.91 [#/sec] (mean)</summary>
 
 ```bash
Server Software:        nginx/1.21.0
Server Hostname:        stern-nl.localhost
Server Port:            80

Document Path:          /my-uri
Document Length:        348161 bytes

Concurrency Level:      1
Time taken for tests:   30.322 seconds
Complete requests:      50
Failed requests:        0
Total transferred:      17429472 bytes
HTML transferred:       17407722 bytes
Requests per second:    1.65 [#/sec] (mean)
Time per request:       606.450 [ms] (mean)
Time per request:       606.450 [ms] (mean, across all concurrent requests)
Transfer rate:          561.33 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:   504  606  88.9    571     789
Waiting:      503  605  88.8    570     789
Total:        504  606  88.9    571     789

Percentage of the requests served within a certain time (ms)
  50%    571
  66%    676
  75%    693
  80%    700
  90%    729
  95%    774
  98%    789
  99%    789
 100%    789 (longest request)
 ```
 
 </details>

  - https://blackfire.io/profiles/30ffb1b2-540d-4a7b-8831-68d3fe6c4564/graph?settings%5Bdimension%5D=wt&settings%5Bdisplay%5D=landscape&settings%5BtabPane%5D=nodes&selected=&callname=main()
 
<img width="626" alt="Bildschirmfoto 2021-06-01 um 01 35 43" src="https://user-images.githubusercontent.com/1698337/120249041-cfea4400-c279-11eb-9238-6aec0a881ae9.png">
 
----
 
 This PR tries to use an more efficient way of parsing the xml structure sadly in costs of readability of the code, hope I can get into a form which it would be still maintainable.
 
 The PR is not yet finished and I'm just experimenting here a little bit. The blackfire profiles seems currently not show the result I'm seeing in the response times, need to investigate here more time. Update added the apache bench benchmark results for better performance testing.