@alexander-schranz 2021-06-01
@dbu what do you think in general about changing the xml parsing here?