@alexander-schranz 2021-06-06
@dbu Did now create a class and use it inside `xmlToProps` and `xmlToColumns` methods. Found some edge cases I didn't take into account (sv:property without sv:value, concated data values when having encoded string in the xml (foo &amp; bar). Basically they did fail in the whole test suit, still I created additional unit test for the new class as it was easier for testing. Pull Request is from my side now ready for review.