@dbu 2021-06-07
https://github.com/jackalope/jackalope-doctrine-dbal/releases/tag/1.7.0