@plozmun 2021-06-07
Hi @alexander-schranz , i've tested this change and it is a great improvement. Thank you!!! 👏