@alexander-schranz 2021-06-07
@plozmun thank you for testing this out. I'm curious how much the performance difference it is in your case for uncached pages in prod env.