@plozmun 2021-06-07
> @plozmun thank you for testing this out. I'm curious how much the performance difference it is in your case for uncached pages in prod env.

Frontend performance has improved over 70-80% for uncached pages

In the admin is where the improvement is huge, from having some timeout to responses in few seconds for sections with many pages.