@alexander-schranz 2021-06-07
@plozmun Wow 70-80% this is really great, would not have expected that 😲. So your timeouts issues are now gone? Which sulu version did you use to test it?