@plozmun 2021-06-07
@alexander-schranz yes! We have pages with many relations and 20 languages and those pages are where it really improves. Sulu version is 2.3