@alexander-schranz 2021-06-07
@plozmun Really great to hear that this did also solve your performance issues. Think with this stats you did use the `2.3@dev` version wich also includes https://github.com/sulu/sulu/pull/6089 as this also improve the performance when loading nodes.
Thank you for your feedback and testing this so out so early 🙏.